requirements.txt
keyboard
opencv-python
pyautogui
numpy
mss
Pillow
label-studio
pygame